% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_groupcond_logl}
\alias{compute_groupcond_logl}
\title{Compute the group-conditional log-likelihood function}
\usage{
compute_groupcond_logl(x, muy, sigmayw, muwz, sigmaz, ghweight, detvarz, dims)
}
\arguments{
\item{x}{Numeric matrix}
}
\description{
Compute the group-conditional log-likelihood function
}
