% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countreg.R
\name{countreg}
\alias{countreg}
\title{Fitting Count Regression Models with Latent Covariates}
\usage{
countreg(formula, lv, group, data, family)
}
\arguments{
\item{formula}{an object of class for \code{\link[stats]{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under Details.}

\item{lv}{Definition of the latent variables.}

\item{group}{A group variable}

\item{data}{a data frame}

\item{family}{Poisson or negative binomial}
}
\value{
An object of type \code{countreg}.
}
\description{
Description
}
\details{
Details
}
\examples{
add(1, 1)
add(10, 1)

}
