% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creg_cfa.r
\name{creg_cfa}
\alias{creg_cfa}
\title{CFA function as a wrapper for countreg}
\usage{
creg_cfa(
  lv,
  data,
  group = NULL,
  silent = FALSE,
  se = FALSE,
  creg_options = NULL
)
}
\description{
CFA function as a wrapper for countreg
}
