% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creg_create_input.R
\name{creg_create_input}
\alias{creg_create_input}
\title{Create input}
\usage{
creg_create_input(forml, lv, group, data, family, silent, se)
}
\arguments{
\item{forml}{an object of class for \code{\link[stats]{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under Details.}

\item{lv}{Definition of the latent variables.}

\item{group}{A group variable}

\item{data}{a data frame}

\item{family}{Poisson or negative binomial}

\item{silent}{Should informations about the estimation process be suppressed?}

\item{se}{Should standard errors be computed? (Can take a while for complex models)}
}
\description{
Turns the input into an input object
}
\keyword{internal}
