% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creg_loglikelihood.R
\name{creg_fit_model}
\alias{creg_fit_model}
\title{Fit the lavacreg model}
\usage{
creg_fit_model(object)
}
\description{
A wrapper for starting values, optimizing loglik and computation of standard errors
}
\details{
@param object A lavacreg object
}
\keyword{internal}
