% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creg_loglikelihood.R
\name{creg_loglikelihood_function}
\alias{creg_loglikelihood_function}
\title{lavacreg log-likelihood function}
\usage{
creg_loglikelihood_function(datalist, modellist)
}
\arguments{
\item{datalist}{Datalist}

\item{modellist}{Modellist}
}
\description{
Computes the likelihood function using the C++ function for group-conditional likelihoods
}
\keyword{internal}
