% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creg_partable.R
\name{creg_partable}
\alias{creg_partable}
\title{Create An Initial Partable for Count Regression Model}
\usage{
creg_partable(object)
}
\arguments{
\item{object}{An object of class \code{countReg}, which contains input and datalist at this point}
}
\description{
The partable serves three purposes:
1. Translating the input into the required parameters
2. The initial \code{par} column contains starting values for each free parameter
3. Holds the values of each iteration within the fitting process and connects them to model
}
\details{
Details
}
