% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{example01}
\alias{example01}
\title{A first example to illustrate}
\format{
A data frame with 871 rows and 9 variables:
\describe{
  \item{dv}{Count of correctly-answered items (dependent variable)}
  \item{treat}{Treatment group variable, where \code{0} is control and \code{2} is treatment}
  \item{z11-z12}{Indicators of LoC}
  \item{z21-z22}{Indicators of LoC}
  \item{z41-z43}{Indicators of depression}
}
}
\usage{
example01
}
\description{
An ACTIVE dataset containing ... .
}
\keyword{datasets}
