% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creg_create_input.R
\name{is.count}
\alias{is.count}
\title{Count variable check}
\usage{
is.count(x, tol = .Machine$double.eps^0.5)
}
\description{
Checks if the variable is a count variable
}
\details{
@param x vector to be checked
 @param tol Tolerance
}
