% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavacreg-package.R
\docType{package}
\name{lavacreg-package}
\alias{lavacreg}
\alias{lavacreg-package}
\title{lavacreg: Latent Variable Count Regression Models}
\description{
Estimation of a multi-group count regression models (i.e., Poisson, negative binomial) with latent covariates. This packages provides two extensions compared to ordinary count regression models based on a generalized linear model: First, measurement models for the predictors can be specified allowing to account for measurement error. Second, the count regression can be simultaneously estimated in multiple groups with stochastic group weights. The marginal maximum likelihood estimation is described in Kiefer & Mayer (2020) \doi{10.1080/00273171.2020.1751027}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chkiefer/lavacreg}
  \item Report bugs at \url{https://github.com/chkiefer/lavacreg/issues}
}

}
\author{
\strong{Maintainer}: Christoph Kiefer \email{christoph.kiefer@uni-bielefeld.de} (\href{https://orcid.org/0000-0002-9166-400X}{ORCID})

}
\keyword{internal}
