% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00methods.R
\name{summary,lavacreg-method}
\alias{summary,lavacreg-method}
\title{Summary of a lavacreg object}
\usage{
\S4method{summary}{lavacreg}(object)
}
\arguments{
\item{object}{lavacreg object}
}
\description{
Summary of a lavacreg object
}
