% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_groupcond_logl}
\alias{compute_groupcond_logl}
\title{Compute the group-conditional log-likelihood function}
\usage{
compute_groupcond_logl(x, muy, sigmayw, muwz, sigmaz, ghweight, detvarz, dims)
}
\arguments{
\item{x}{Model matrix containing the observed variables in order Y, W, Z}

\item{muy}{Matrix of conditional expectations of Y given Z and/or latent covariates}

\item{sigmayw}{Vector containing the size parameter (for overdispersion) and measurement error variances}

\item{muwz}{Matrix of conditional expectations of W and Z given latent covariate (if existent) and given the group}

\item{sigmaz}{(Conditional) covariance matrix of observed covariates Z}

\item{ghweight}{Gauss-Hermite weights, if latent covariates included}

\item{detvarz}{Determinant of sigmaz (easier to compute in R...)}

\item{dims}{Integer vector with dimensional information, in order group size, integration points, number of Z, and number of W}
}
\description{
Compute the group-conditional log-likelihood function
}
