% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{example01}
\alias{example01}
\title{A first example dataset to illustrate the use of lavacreg}
\format{
A data frame with 871 rows and 9 variables:
\describe{
  \item{dv}{Count of correctly-answered items (dependent variable)}
  \item{treat}{Treatment group variable, where \code{0} is control and \code{2} is treatment}
  \item{z11}{First indicator of internal LoC}
  \item{z12}{Second indicator of internal LoC}
  \item{z21}{First indicator of external LoC}
  \item{z22}{Second indicator of external LoC}
  \item{z41}{First indicator of depression}
  \item{z42}{Second indicator of depression}
  \item{z43}{Third indicator of depression}
}
}
\usage{
example01
}
\description{
A dataset containing 9 variables: a dependent variable dv, a group variable treat
and 7 indicators for 3 latent covariates.
}
\keyword{datasets}
