% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00methods.R
\name{summary,lavacreg-method}
\alias{summary,lavacreg-method}
\title{Summary of a lavacreg object}
\usage{
\S4method{summary}{lavacreg}(object)
}
\arguments{
\item{object}{lavacreg object}
}
\value{
Function prints the parameter table of an estimated model, which
includes the parameter estimates and standard errors.
}
\description{
Exports the parameter table with parameter estimates and standard errors 
for an estimated latent variable count regression model.
}
